refactor(bin): move awaiters into same file as Pep/10 CPU simulator
Awaiters are tied to the specific class implemented, so no need to put them in a separate header.
