+++ bin/simexp/CMakeLists.txt
@@ -3,7 +3,6 @@ cmake_minimum_required(VERSION 3.16)
 qt_add_executable(
   simexp
   main.cpp
-  simuloop.cpp
   pep10isa.cpp
   event_dispatch.cpp
   event_queue.cpp
